bug fixes viewed by sonarcloud.

+++ server/src/main/java/org/tango/server/idl/TangoIDLUtil.java
@@ -1,32 +1,28 @@
 /**
  * Copyright (C) :     2012
- *
- * 	Synchrotron Soleil
- * 	L'Orme des merisiers
- * 	Saint Aubin
- * 	BP48
- * 	91192 GIF-SUR-YVETTE CEDEX
- *
+ * <p>
+ * Synchrotron Soleil
+ * L'Orme des merisiers
+ * Saint Aubin
+ * BP48
+ * 91192 GIF-SUR-YVETTE CEDEX
+ * <p>
  * This file is part of Tango.
- *
+ * <p>
  * Tango is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- *
+ * <p>
  * Tango is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU Lesser General Public License for more details.
- *
+ * <p>
  * You should have received a copy of the GNU Lesser General Public License
  * along with Tango.  If not, see <http://www.gnu.org/licenses/>.
  */
 package org.tango.server.idl;
-
-import org.tango.server.pipe.PipeConfiguration;
-import org.tango.server.pipe.PipeImpl;
-import org.tango.server.pipe.PipeValue;
 
 import fr.esrf.Tango.DevFailed;
 import fr.esrf.Tango.DevPipeBlob;
@@ -34,6 +30,9 @@ import fr.esrf.Tango.DevPipeData;
 import fr.esrf.Tango.PipeConfig;
 import fr.esrf.Tango.TimeVal;
 import fr.esrf.TangoApi.PipeBlob;
+import org.tango.server.pipe.PipeConfiguration;
+import org.tango.server.pipe.PipeImpl;
+import org.tango.server.pipe.PipeValue;
 
 public final class TangoIDLUtil {
 
@@ -50,7 +49,7 @@ public final class TangoIDLUtil {
         return timeVal;
     }
 
-    public static long getTime(final TimeVal timeVal) {
+    public static int getTime(final TimeVal timeVal) {
         return timeVal.tv_nsec * FACTOR_1000 + timeVal.tv_usec * FACTOR_1000 * FACTOR_1000;
     }
